fix(ci): keep image admission verification within CI limits
Why: the worker-deployment parent branch failed CI on strict index-signature access, an optional workflow pin value, and the 500-line image script limit. Its new admission paths also reduced migrated-script coverage.\n\nWhat: isolate candidate admission and fallback-report helpers, narrow the optional pin before comparison, use bracket access for catalog records, and cover the admission success and failure paths.\n\nVerification: bun run scripts/check-script-coverage.ts; bun run scripts/lint-buildkite.ts; bunx --no-install tsc --noEmit -p scripts/tsconfig.json; bunx --no-install tsc --noEmit -p packages/homelab/src/cdk8s/tsconfig.json; bun --no-install --bun vitest --config vitest.config.ts run .buildkite/scripts/bake-images.test.ts --no-file-parallelism
